pyproject.toml
[project]
name = "websearch-ai"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "DrinkingPants74", email = "" }
]
dependencies = [
    "flet>=0.28.3",
    "llama-cpp-python",
    "httpx",
    "beautifulsoup4",
    "readability-lxml",
    "pypng",
    "numpy",

   ]

[tool.flet]
# org name in reverse domain name notation, e.g. "com.mycompany".
# Combined with project.name to build bundle ID for iOS and Android apps
org = "com.OneShotEntertainment"

# project display name that is used as an app title on Android and iOS home screens,
# shown in window titles and about app dialogs on desktop.
product = "WebSearch AI"

# company name to display in about app dialogs
company = "One Shot Entertainment"

# copyright text to display in about app dialogs
copyright = "GPLv3"

# [tool.flet.app]
# path = "src"

[tool.uv]
dev-dependencies = [
    "flet[all]==0.28.3",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
flet = {extras = ["all"], version = "0.28.3"}
